+++ src/append.ts
@@ -1,0 +1,6 @@
+export function append(
+  parent: ParentNode,
+  ...childs: Array<string | Node>
+): void {
+  parent.append(...childs);
+}