+++ src/index.ts
@@ -1,9 +1,10 @@
-export { closest } from "./closest";
+export { append } from './append';
+export { closest } from './closest';
 export { setAttribute } from './set_attribute';
 export { getAttribute } from './get_attribute';
 export { hasAttribute } from './has_attribute';
-export { querySelector } from "./query_selector";
-export { createElement } from "./create_element";
-export { createTextNode } from "./create_text_node";
+export { querySelector } from './query_selector';
+export { createElement } from './create_element';
+export { createTextNode } from './create_text_node';
 export { removeAttribute } from './remove_attribute';
-export { querySelectorAll } from "./query_selector_all";
+export { querySelectorAll } from './query_selector_all';